PUBDEV-6447 set constrained kmeans test aside
